Send `DeviceConfigured` MDM command after DEP enrollment (#17737)
