+++ server/worker/apple_mdm_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/fleetdm/fleet/v4/pkg/optjson"
 	"github.com/fleetdm/fleet/v4/server/datastore/mysql"
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	apple_mdm "github.com/fleetdm/fleet/v4/server/mdm/apple"
@@ -40,6 +41,8 @@ func TestAppleMDM(t *testing.T) {
 	// specific internals (sequence and number of calls, etc.). The MDM storage
 	// and pusher are mocks.
 	ds := mysql.CreateMySQLDS(t)
+	// call TruncateTables immediately as a DB migation may have created jobs
+	mysql.TruncateTables(t, ds)
 
 	mdmStorage, err := ds.NewMDMAppleMDMStorage([]byte("test"), []byte("test"))
 	require.NoError(t, err)
@@ -92,6 +95,32 @@ func TestAppleMDM(t *testing.T) {
 		return commands
 	}
 
+	enableManualRelease := func(t *testing.T, teamID *uint) {
+		if teamID == nil {
+			enableAppCfg := func(enable bool) {
+				ac, err := ds.AppConfig(ctx)
+				require.NoError(t, err)
+				ac.MDM.MacOSSetup.EnableReleaseDeviceManually = optjson.SetBool(enable)
+				err = ds.SaveAppConfig(ctx, ac)
+				require.NoError(t, err)
+			}
+
+			enableAppCfg(true)
+			t.Cleanup(func() { enableAppCfg(false) })
+		} else {
+			enableTm := func(enable bool) {
+				tm, err := ds.Team(ctx, *teamID)
+				require.NoError(t, err)
+				tm.Config.MDM.MacOSSetup.EnableReleaseDeviceManually = optjson.SetBool(enable)
+				_, err = ds.SaveTeam(ctx, tm)
+				require.NoError(t, err)
+			}
+
+			enableTm(true)
+			t.Cleanup(func() { enableTm(false) })
+		}
+	}
+
 	t.Run("no-op with nil commander", func(t *testing.T) {
 		defer mysql.TruncateTables(t, ds)
 
@@ -115,7 +144,7 @@ func TestAppleMDM(t *testing.T) {
 		// again
 		time.Sleep(time.Second)
 
-		jobs, err := ds.GetQueuedJobs(ctx, 1)
+		jobs, err := ds.GetQueuedJobs(ctx, 1, time.Now().UTC().Add(time.Minute)) // look in the future to catch any delayed job
 		require.NoError(t, err)
 		require.Empty(t, jobs)
 	})
@@ -143,7 +172,7 @@ func TestAppleMDM(t *testing.T) {
 		// ensure the job's not_before allows it to be returned
 		time.Sleep(time.Second)
 
-		jobs, err := ds.GetQueuedJobs(ctx, 1)
+		jobs, err := ds.GetQueuedJobs(ctx, 1, time.Time{})
 		require.NoError(t, err)
 		require.Len(t, jobs, 1)
 		require.Contains(t, jobs[0].Error, "unknown task: no-such-task")
@@ -175,9 +204,49 @@ func TestAppleMDM(t *testing.T) {
 		// again
 		time.Sleep(time.Second)
 
-		jobs, err := ds.GetQueuedJobs(ctx, 1)
+		jobs, err := ds.GetQueuedJobs(ctx, 1, time.Now().UTC().Add(time.Minute)) // look in the future to catch any delayed job
 		require.NoError(t, err)
+
+		// the post-DEP release device job is pending
+		require.Len(t, jobs, 1)
+		require.Equal(t, appleMDMJobName, jobs[0].Name)
+		require.Contains(t, string(*jobs[0].Args), AppleMDMPostDEPReleaseDeviceTask)
+		require.Equal(t, 0, jobs[0].Retries) // hasn't run yet
+
+		require.ElementsMatch(t, []string{"InstallEnterpriseApplication"}, getEnqueuedCommandTypes(t))
+	})
+
+	t.Run("installs default manifest, manual release", func(t *testing.T) {
+		t.Cleanup(func() { mysql.TruncateTables(t, ds) })
+
+		h := createEnrolledHost(t, 1, nil, true)
+		enableManualRelease(t, nil)
+
+		mdmWorker := &AppleMDM{
+			Datastore: ds,
+			Log:       nopLog,
+			Commander: apple_mdm.NewMDMAppleCommander(mdmStorage, mockPusher{}),
+		}
+		w := NewWorker(ds, nopLog)
+		w.Register(mdmWorker)
+
+		err = QueueAppleMDMJob(ctx, ds, nopLog, AppleMDMPostDEPEnrollmentTask, h.UUID, nil, "")
+		require.NoError(t, err)
+
+		// run the worker, should succeed
+		err = w.ProcessJobs(ctx)
+		require.NoError(t, err)
+
+		// ensure the job's not_before allows it to be returned if it were to run
+		// again
+		time.Sleep(time.Second)
+
+		jobs, err := ds.GetQueuedJobs(ctx, 1, time.Now().UTC().Add(time.Minute)) // look in the future to catch any delayed job
+		require.NoError(t, err)
+
+		// there is no post-DEP release device job pending
 		require.Empty(t, jobs)
+
 		require.ElementsMatch(t, []string{"InstallEnterpriseApplication"}, getEnqueuedCommandTypes(t))
 	})
 
@@ -213,9 +282,15 @@ func TestAppleMDM(t *testing.T) {
 		// again
 		time.Sleep(time.Second)
 
-		jobs, err := ds.GetQueuedJobs(ctx, 1)
+		jobs, err := ds.GetQueuedJobs(ctx, 1, time.Now().UTC().Add(time.Minute)) // look in the future to catch any delayed job
 		require.NoError(t, err)
-		require.Empty(t, jobs)
+
+		// the post-DEP release device job is pending
+		require.Len(t, jobs, 1)
+		require.Equal(t, appleMDMJobName, jobs[0].Name)
+		require.Contains(t, string(*jobs[0].Args), AppleMDMPostDEPReleaseDeviceTask)
+		require.Equal(t, 0, jobs[0].Retries) // hasn't run yet
+
 		require.ElementsMatch(t, []string{"InstallEnterpriseApplication", "InstallEnterpriseApplication"}, getEnqueuedCommandTypes(t))
 
 		ms, err := ds.GetHostMDMMacOSSetup(ctx, h.ID)
@@ -258,9 +333,64 @@ func TestAppleMDM(t *testing.T) {
 		// again
 		time.Sleep(time.Second)
 
-		jobs, err := ds.GetQueuedJobs(ctx, 1)
+		jobs, err := ds.GetQueuedJobs(ctx, 1, time.Now().UTC().Add(time.Minute)) // look in the future to catch any delayed job
 		require.NoError(t, err)
+
+		// the post-DEP release device job is pending
+		require.Len(t, jobs, 1)
+		require.Equal(t, appleMDMJobName, jobs[0].Name)
+		require.Contains(t, string(*jobs[0].Args), AppleMDMPostDEPReleaseDeviceTask)
+		require.Equal(t, 0, jobs[0].Retries) // hasn't run yet
+
+		require.ElementsMatch(t, []string{"InstallEnterpriseApplication", "InstallEnterpriseApplication"}, getEnqueuedCommandTypes(t))
+
+		ms, err := ds.GetHostMDMMacOSSetup(ctx, h.ID)
+		require.NoError(t, err)
+		require.Equal(t, "custom-team-bootstrap", ms.BootstrapPackageName)
+	})
+
+	t.Run("installs custom bootstrap manifest of a team, manual release", func(t *testing.T) {
+		t.Cleanup(func() { mysql.TruncateTables(t, ds) })
+
+		tm, err := ds.NewTeam(ctx, &fleet.Team{Name: "test"})
+		require.NoError(t, err)
+		enableManualRelease(t, &tm.ID)
+
+		h := createEnrolledHost(t, 1, &tm.ID, true)
+		err = ds.InsertMDMAppleBootstrapPackage(ctx, &fleet.MDMAppleBootstrapPackage{
+			Name:   "custom-team-bootstrap",
+			TeamID: tm.ID,
+			Bytes:  []byte("test"),
+			Sha256: []byte("test"),
+			Token:  "token",
+		})
+		require.NoError(t, err)
+
+		mdmWorker := &AppleMDM{
+			Datastore: ds,
+			Log:       nopLog,
+			Commander: apple_mdm.NewMDMAppleCommander(mdmStorage, mockPusher{}),
+		}
+		w := NewWorker(ds, nopLog)
+		w.Register(mdmWorker)
+
+		err = QueueAppleMDMJob(ctx, ds, nopLog, AppleMDMPostDEPEnrollmentTask, h.UUID, &tm.ID, "")
+		require.NoError(t, err)
+
+		// run the worker, should succeed
+		err = w.ProcessJobs(ctx)
+		require.NoError(t, err)
+
+		// ensure the job's not_before allows it to be returned if it were to run
+		// again
+		time.Sleep(time.Second)
+
+		jobs, err := ds.GetQueuedJobs(ctx, 1, time.Now().UTC().Add(time.Minute)) // look in the future to catch any delayed job
+		require.NoError(t, err)
+
+		// there is no post-DEP release device job pending
 		require.Empty(t, jobs)
+
 		require.ElementsMatch(t, []string{"InstallEnterpriseApplication", "InstallEnterpriseApplication"}, getEnqueuedCommandTypes(t))
 
 		ms, err := ds.GetHostMDMMacOSSetup(ctx, h.ID)
@@ -292,7 +422,7 @@ func TestAppleMDM(t *testing.T) {
 		// again
 		time.Sleep(time.Second)
 
-		jobs, err := ds.GetQueuedJobs(ctx, 1)
+		jobs, err := ds.GetQueuedJobs(ctx, 1, time.Time{})
 		require.NoError(t, err)
 		require.Len(t, jobs, 1)
 		require.Contains(t, jobs[0].Error, "MDMIdPAccount with uuid abcd was not found")
@@ -334,9 +464,15 @@ func TestAppleMDM(t *testing.T) {
 		// again
 		time.Sleep(time.Second)
 
-		jobs, err := ds.GetQueuedJobs(ctx, 1)
+		jobs, err := ds.GetQueuedJobs(ctx, 1, time.Now().UTC().Add(time.Minute)) // look in the future to catch any delayed job
 		require.NoError(t, err)
-		require.Empty(t, jobs)
+
+		// the post-DEP release device job is pending, having failed its first attempt
+		require.Len(t, jobs, 1)
+		require.Equal(t, appleMDMJobName, jobs[0].Name)
+		require.Contains(t, string(*jobs[0].Args), AppleMDMPostDEPReleaseDeviceTask)
+		require.Equal(t, 0, jobs[0].Retries) // hasn't run yet
+
 		// confirm that AccountConfiguration command was not enqueued
 		require.ElementsMatch(t, []string{"InstallEnterpriseApplication"}, getEnqueuedCommandTypes(t))
 	})
@@ -383,9 +519,15 @@ func TestAppleMDM(t *testing.T) {
 		// again
 		time.Sleep(time.Second)
 
-		jobs, err := ds.GetQueuedJobs(ctx, 1)
+		jobs, err := ds.GetQueuedJobs(ctx, 1, time.Now().UTC().Add(time.Minute)) // look in the future to catch any delayed job
 		require.NoError(t, err)
-		require.Empty(t, jobs)
+
+		// the post-DEP release device job is pending
+		require.Len(t, jobs, 1)
+		require.Equal(t, appleMDMJobName, jobs[0].Name)
+		require.Contains(t, string(*jobs[0].Args), AppleMDMPostDEPReleaseDeviceTask)
+		require.Equal(t, 0, jobs[0].Retries) // hasn't run yet
+
 		require.ElementsMatch(t, []string{"InstallEnterpriseApplication", "AccountConfiguration"}, getEnqueuedCommandTypes(t))
 	})
 
@@ -413,7 +555,7 @@ func TestAppleMDM(t *testing.T) {
 		// again
 		time.Sleep(time.Second)
 
-		jobs, err := ds.GetQueuedJobs(ctx, 1)
+		jobs, err := ds.GetQueuedJobs(ctx, 1, time.Now().UTC().Add(time.Minute)) // look in the future to catch any delayed job
 		require.NoError(t, err)
 		require.Empty(t, jobs)
 		require.ElementsMatch(t, []string{"InstallEnterpriseApplication"}, getEnqueuedCommandTypes(t))
