+++ server/worker/macos_setup_assistant_test.go
@@ -25,6 +25,8 @@ import (
 func TestMacosSetupAssistant(t *testing.T) {
 	ctx := context.Background()
 	ds := mysql.CreateMySQLDS(t)
+	// call TruncateTables immediately as some DB migrations may create jobs
+	mysql.TruncateTables(t, ds)
 
 	// create a couple hosts for no team, team 1 and team 2 (none for team 3)
 	hosts := make([]*fleet.Host, 6)
@@ -140,7 +142,7 @@ func TestMacosSetupAssistant(t *testing.T) {
 		err = w.ProcessJobs(ctx)
 		require.NoError(t, err)
 		// no remaining jobs to process
-		pending, err := ds.GetQueuedJobs(ctx, 10)
+		pending, err := ds.GetQueuedJobs(ctx, 10, time.Time{})
 		require.NoError(t, err)
 		require.Empty(t, pending)
 	}
